+++ bitalarm-impl/src/main/java/com/dwarfeng/bitalarm/impl/handler/pusher/DrainPusher.java
@@ -1,27 +1,24 @@
 package com.dwarfeng.bitalarm.impl.handler.pusher;
 
-import com.dwarfeng.bitalarm.impl.handler.Pusher;
 import com.dwarfeng.bitalarm.stack.bean.entity.AlarmHistory;
 import com.dwarfeng.bitalarm.stack.bean.entity.AlarmInfo;
 import org.springframework.stereotype.Component;
 
 import java.util.List;
-import java.util.Objects;
 
 /**
  * 简单的丢弃掉所有信息的推送器。
  *
  * @author DwArFeng
- * @since 1.0.0
+ * @since 1.5.3
  */
 @Component
-public class DrainPusher implements Pusher {
+public class DrainPusher extends AbstractPusher {
 
-    public static final String SUPPORT_TYPE = "drain";
+    public static final String PUSHER_TYPE = "drain";
 
-    @Override
-    public boolean supportType(String type) {
-        return Objects.equals(SUPPORT_TYPE, type);
+    public DrainPusher() {
+        super(PUSHER_TYPE);
     }
 
     @Override