+++ bitalarm-impl/src/main/java/com/dwarfeng/bitalarm/impl/handler/pusher/MultiPusher.java
@@ -13,7 +13,6 @@ import org.springframework.stereotype.Component;
 import javax.annotation.PostConstruct;
 import java.util.ArrayList;
 import java.util.List;
-import java.util.Objects;
 import java.util.StringTokenizer;
 
 /**
@@ -23,9 +22,10 @@ import java.util.StringTokenizer;
  * @since 1.0.0
  */
 @Component
-public class MultiPusher implements Pusher {
+public class MultiPusher extends AbstractPusher {
 
-    public static final String SUPPORT_TYPE = "multi";
+    public static final String PUSHER_TYPE = "multi";
+
     private static final Logger LOGGER = LoggerFactory.getLogger(MultiPusher.class);
 
     @Autowired
@@ -36,6 +36,10 @@ public class MultiPusher implements Pusher {
 
     private final List<Pusher> delegates = new ArrayList<>();
 
+    public MultiPusher() {
+        super(PUSHER_TYPE);
+    }
+
     @PostConstruct
     public void init() throws HandlerException {
         StringTokenizer st = new StringTokenizer(delegateTypes, ",");
@@ -44,11 +48,6 @@ public class MultiPusher implements Pusher {
             delegates.add(pushers.stream().filter(p -> p.supportType(delegateType)).findAny()
                     .orElseThrow(() -> new HandlerException("未知的 pusher 类型: " + delegateType)));
         }
-    }
-
-    @Override
-    public boolean supportType(String type) {
-        return Objects.equals(SUPPORT_TYPE, type);
     }
 
     @Override